backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt==4.0.1",
    "fal-client>=0.8.1",
    "fastapi>=0.119.0",
    "google-genai>=1.45.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.12.3",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pdoc>=15.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest-cov>=6.0.0",
    "types-passlib>=1.7.7.20241221",
]
